fix(receipts): move buildQueueItems off the client boundary
The server-render crash on /receipts/review was: server components
imported buildQueueItems() from queue-rail.tsx, which is a "use client"
module. Next.js wraps non-component exports from client modules as client
references, so calling them on the server throws "Attempted to call X
from the server but X is on the client."

Extract buildQueueItems + QueueItem + the pure formatters into
lib/receipts/queue-items.ts (no "use client") and update the two server
pages to import from there. QueueRail itself stays client-side.
