+++ app/(receipt-system)/receipts/review/[id]/page.tsx
@@ -6,10 +6,8 @@ import {
 } from "@/lib/receipts/db";
 import { assertReceiptsPageAccess } from "@/lib/receipts/auth-request";
 import { ReviewLayout } from "@/components/receipts/review/review-layout";
-import {
-  QueueRail,
-  buildQueueItems,
-} from "@/components/receipts/review/queue-rail";
+import { QueueRail } from "@/components/receipts/review/queue-rail";
+import { buildQueueItems } from "@/lib/receipts/queue-items";
 import { ImagePane } from "@/components/receipts/review/image-pane";
 import { FormPane } from "@/components/receipts/review/form-pane";
 import {
